requirements.txt
jupyterlab==4.3.5
numpy==2.2.2
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
scipy==1.15.1
scikit-learn==1.6.1
fsspec==2025.5.1
gcsfs==2025.5.1
xgboost==3.0.2
pyspark==3.5.5
tqdm==4.67.1
google-cloud-storage
fastapi==0.110.2
uvicorn[standard]==0.29.0
joblib==1.4.2
apache-airflow[celery,postgres,redis]==2.9.2
psycopg2-binary
flower